+++ src/genetic_algorithm/genetic_algorithm_methods.h
@@ -1,4 +1,4 @@
-float fitnessFunction(float timeBeforeCrash, float distanceTraveledBeforeCrash);
+double fitnessFunction(double timeBeforeCrash, double distanceTraveledBeforeCrash);
 vector<DNA> pickBestDNAs(vector<DNA> DNAs);
 DNA crossover(vector<DNA> bestDNAs);
 vector<DNA> mutation(DNA dna, int N_CARS);
